feat: implement lotto result and print result
당첨 로또 검증 구현
당첨 로또 출력
로또 수익률 출력

+++ src/main/java/lotto/LottoMachine.java
@@ -8,7 +8,7 @@ public class LottoMachine {
 
 
     public void sellLottoTo(Customer customer) {
-        int count = customer.choseLottoCount();
+        Long count = customer.choseLottoCount();
         for (int i = 0; i <count; i++) {
             customer.buy(makeLotto());
         }
